Закончил работу с адресами
Добавил поиск адреса - get запрос и получение цепочки адресов - get запрос, добавил на полчение цепочки проверку на исключение, что пользователь ввел коректный адрес

+++ DeliveryFoodBackend/DTO/SearchAddressModel.cs
@@ -6,7 +6,7 @@ namespace DeliveryFoodBackend.DTO
     public class SearchAddressModel
     {
         [Required]
-        public int ObjectId { get; set; }
+        public long ObjectId { get; set; }
 
         [Required]
         public Guid ObgectGuid { get; set; }
